+++ src/Products/PloneMeeting/columns.py
@@ -250,6 +250,20 @@ class PMActionsColumn(ActionsColumn):
         return super(ActionsColumn, self).renderCell(item)
 
 
+class PMAsyncActionsColumn(BaseColumn):
+    """A column that displays an icon that will on hover display a tooltipster
+       with actions availble to current user."""
+
+    sort_index = -1
+
+    def renderCell(self, item):
+        tag = """<div class="async-actions-panel-icon-container">""" \
+        """<a href="#" title="Actions" onclick="event.preventDefault();;">""" \
+        """<span class="fa async-actions-panel-icon link-action tooltipster-actions-panel" data-base_url="{}" """ \
+        """data-showHistory:boolean="1" /></a></div>""".format(item.getURL())
+        return tag
+
+
 class ItemLinkedMeetingColumn(BaseColumn):
     """
       Display the formatted date and a link to the linked meeting if any.