+++ src/Products/PloneMeeting/tests/testContacts.py
@@ -27,7 +27,6 @@ from Products.PloneMeeting.Extensions.imports import import_contacts
 from Products.PloneMeeting.tests.PloneMeetingTestCase import PloneMeetingTestCase
 from Products.statusmessages.interfaces import IStatusMessage
 from z3c.relationfield.relation import RelationValue
-from zExceptions import Redirect
 from zope.component import getUtility
 from zope.event import notify
 from zope.i18n import translate
@@ -115,14 +114,14 @@ class testContacts(PloneMeetingTestCase):
         self.assertTrue(hp_uid in meeting.get_attendees())
         # hp not deletable because used in MC and meeting
         self.changeUser('siteadmin')
-        self.assertRaises(Redirect, api.content.delete, hp)
+        self.assertRaises(BeforeDeleteException, api.content.delete, hp)
 
         # unselect from MeetingConfig.orderedContacts,
         # still not deletable because used by a meeting
         orderedContacts = list(cfg.getOrderedContacts())
         orderedContacts.remove(hp_uid)
         cfg.setOrderedContacts(orderedContacts)
-        self.assertRaises(Redirect, api.content.delete, hp)
+        self.assertRaises(BeforeDeleteException, api.content.delete, hp)
 
         # unselect hp from meeting, now it is deletable
         del meeting.ordered_contacts[hp.UID()]
@@ -138,14 +137,14 @@ class testContacts(PloneMeetingTestCase):
 
         # hp not deletable because used in MC and meeting item
         self.changeUser('siteadmin')
-        self.assertRaises(Redirect, api.content.delete, hp)
+        self.assertRaises(BeforeDeleteException, api.content.delete, hp)
 
         # unselect from MeetingConfig.orderedItemInitiators,
         # still not deletable because used by a meeting item
         orderedItemInitiators = list(cfg.getOrderedItemInitiators())
         orderedItemInitiators.remove(hp_uid)
         cfg.setOrderedItemInitiators(orderedItemInitiators)
-        self.assertRaises(Redirect, api.content.delete, hp)
+        self.assertRaises(BeforeDeleteException, api.content.delete, hp)
 
         # unselect hp from meeting item, now it is deletable
         item.setItemInitiator(())
@@ -1794,6 +1793,7 @@ class testContacts(PloneMeetingTestCase):
     def test_pm_RedefinedCertifiedSignatures(self):
         """organization.certified_signatures may override what is defined on a MeetingConfig,
            either partially (one signature, the other is taken from MeetingConfig) or completely."""
+        self.changeUser('siteadmin')
         cfg = self.meetingConfig
         certified = [
             {'signatureNumber': '1',
@@ -2310,21 +2310,30 @@ class testContacts(PloneMeetingTestCase):
     def test_pm_HeldPositionBackRefs(self):
         """This will display back references on a held_position
            to see where it is used."""
-        self.changeUser('pmManager')
-        meeting = self.create('Meeting')
+        self._enableField("itemInitiator")
         person = self.portal.contacts.get('person1')
         hp = person.get_held_positions()[0]
         hp_uid = hp.UID()
+
+        self.changeUser('pmManager')
+        item = self.create('MeetingItem', itemInitiator=(hp_uid, ))
+        meeting = self.create('Meeting')
         self.assertTrue(hp_uid in meeting.get_attendees())
         viewlet = self._get_viewlet(
             context=hp,
             manager_name='plone.belowcontentbody',
             viewlet_name='held_position_back_references')
-        rendered = viewlet.render()
+        rendered_viewlet = viewlet.render()
+        self.assertTrue(hp_uid in rendered_viewlet)
+        # back refs are loaded asynch by @@load_held_position_back_refs
+        view = hp.restrictedTraverse('@@load_held_position_back_refs')
+        rendered_view = view()
         # used in MeetingConfig
-        self.assertTrue(self.meetingConfig.absolute_url() in rendered)
+        self.assertTrue(self.meetingConfig.absolute_url() in rendered_view)
         # used in meeting
-        self.assertTrue(meeting.absolute_url() in rendered)
+        self.assertTrue(meeting.absolute_url() in rendered_view)
+        # used in item
+        self.assertTrue(item.absolute_url() in rendered_view)
 
     def test_pm_RedefineAttendeePositionForm(self):
         """Test the @@item_redefine_attendee_position_form and